+++ src/ingestion_code/model.py
@@ -1,8 +1,7 @@
 import logging
 
+from config import settings
 from sentence_transformers import SentenceTransformer
-
-from ingestion_code.config import settings
 
 logger = logging.getLogger(__name__)
 
@@ -17,9 +16,5 @@ def get_sentence_transformers_model() -> SentenceTransformer:
         logger.info("Model %s loaded successfully.", settings.LOCAL_MODEL_NAME)
         return model
     except Exception as e:
-        logger.exception(
-            "Failed to load SentenceTransformer model %s", settings.LOCAL_MODEL_NAME
-        )
-        raise RuntimeError(
-            f"Error loading embedding model '{settings.LOCAL_MODEL_NAME}'"
-        ) from e
+        logger.exception("Failed to load SentenceTransformer model %s", settings.LOCAL_MODEL_NAME)
+        raise RuntimeError(f"Error loading embedding model '{settings.LOCAL_MODEL_NAME}'") from e
